service: split consolidation and promotion into their own files
